Fix issue: Get double value from NSNumber, result is incorrect
Refactor code.
How to fix: Use NSDecimalNumber directly instead of double value.

+++ MozoSDK/Classes/Modules/PaymentQR/User Interface/View/PaymentQRViewController.swift
@@ -66,7 +66,7 @@ class PaymentQRViewController: MozoBasicViewController {
             
             lbAmountToSend.text = displayItem.amount.roundAndAddCommas()
             
-            lbAmountEx.text = DisplayUtils.getExchangeTextFromAmount(displayItem.amount)
+            lbAmountEx.text = DisplayUtils.getExchangeTextFromAmount(displayItem.amount.doubleValue)
         }
     }
